pyproject.toml
[tool.poetry]
name = "tech-track"
version = "0.1.0"
description = "TechTrack develops a platform for monitoring and managing technical equipment such as industrial robots, manufacturing equipment, quality control systems, etc."
authors = ["Altynai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
django-redis = "^5.4.0"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
